Feat:create base service
