+++ app/services/forecast_service.rb
@@ -1,7 +1,7 @@
-class ForecastService
+class ForecastService < BaseService
   class << self
     def get_all_forecast(location)
-    end_point = "/geocoding/v1/address"
+      end_point = "/geocoding/v1/address"
       response = connection_mapquest.get(end_point) do |faraday|
         faraday.params['key'] = ENV["MAPQUEST_API_KEY"]
         faraday.params["location"] = location
@@ -15,7 +15,7 @@ class ForecastService
     end
 
     def get_weather(latitude, longitude)
-    end_point = "/data/2.5/onecall"
+      end_point = "/data/2.5/onecall"
       response = connection_openweather.get(end_point) do |faraday|
         faraday.params["appid"] = ENV["FORECAST_API_KEY"]
         faraday.params["lat"] = latitude
@@ -24,17 +24,7 @@ class ForecastService
 
       end
       json = JSON.parse(response.body, symbolize_names: true)
-      require "pry"; binding.pry
-    end
-
-    def connection_mapquest
-      url = "http://www.mapquestapi.com"
-      Faraday.new(url: url)
-    end
-
-    def connection_openweather
-      url = "https://api.openweathermap.org"
-      Faraday.new(url: url)
+      # require "pry"; binding.pry
     end
   end
 end
